pyproject.toml
[project]
name = "unfold-reproduction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "django-unfold>=0.76.0",
]
